+++ Sources/LCP/LCPService.swift
@@ -23,13 +23,16 @@ public final class LCPService: Loggable {
 
     /// - Parameter deviceName: Device name used when registering a license to an LSD server.
     ///   If not provided, the device name will be the default `UIDevice.current.name`.
+    /// - Parameter deviceId: Device ID used when registering a license to an LSD server.
+    ///   If not provided, the device id will be generated as a random UUID.
     public init(
         client: LCPClient,
         licenseRepository: LCPLicenseRepository,
         passphraseRepository: LCPPassphraseRepository,
         assetRetriever: AssetRetriever,
         httpClient: HTTPClient,
-        deviceName: String? = nil
+        deviceName: String? = nil,
+        deviceId: String? = nil
     ) {
         // Determine whether the embedded liblcp.a is in production mode, by attempting to open a production license.
         let isProduction: Bool = {
@@ -50,6 +53,7 @@ public final class LCPService: Loggable {
             crl: CRLService(httpClient: httpClient),
             device: DeviceService(
                 deviceName: deviceName ?? UIDevice.current.name,
+                deviceId: deviceId,
                 repository: licenseRepository,
                 httpClient: httpClient
             ),